Move helpers functions into separate file
